requirements.txt
# Core LLM providers
anthropic>=0.40.0
openai>=1.50.0

# AWS SDK
boto3>=1.34.0

# Rich terminal UI
rich>=13.7.0

# Environment
python-dotenv>=1.0.0